refactor setup test logic
